:poop: Allow using template xarray.DataArray to rasterize vector shapes
Convenient but complicated way of rasterizing vector data into a template xarray.DataArray. The vector and raster IterDataPipe objects are Zipped, with the vector IterDataPipe being broadcast/repeated to match the length of the raster IterDataPipe. Affine transformation is obtained directly from xarray.Datarray via the `.rio.transform()` accessor property whenever possible. Automatic reprojection to the xarray.DataArray's coordinate reference system is done if needed (and assuming both have a `crs` attribute). The unit test feels like the length of an integration test, but it captures the key functionalities needed and uses a real-ish dataset.