:sparkles: XarrayCanvasIterDataPipe for creating blank datashader canvas
An iterable-style DataPipe for creating a blank canvas from raster images! The datashader.Canvas is created based on the input xarray object's spatial extent and coordinates, with the metadata obtained using rioxarray. Included a doctest for getting a datashader.Canvas from an xarray.DataArray. Added a new section in the API docs and intersphinx mappings for datashader.