:sparkles: XarrayCanvasIterDataPipe for creating blank datashader canvas (#34)
An iterable-style DataPipe for creating a blank canvas from raster images. This initiates the integration of zen3geo with datashader!

* :heavy_plus_sign: Add datashader and spatialpandas

Quickly and accurately render even the largest data, and pandas extension arrays for spatial/geometric operations!

Unfortunately, had to pin the maximum Python version to 3.11 because scipy (the version with Python 3.10 wheels) sets a hard upper pin for Python <3.12.

* :sparkles: XarrayCanvasIterDataPipe for creating blank datashader canvas

An iterable-style DataPipe for creating a blank canvas from raster images! The datashader.Canvas is created based on the input xarray object's spatial extent and coordinates, with the metadata obtained using rioxarray. Included a doctest for getting a datashader.Canvas from an xarray.DataArray. Added a new section in the API docs and intersphinx mappings for datashader.

* :green_heart: Install datashader and spatialpandas in CI 

Include the 'spatial' extras dependencies in the Continuous Integration tests, and install datashader for Readthedocs. Also mentioned installing the 'spatial' extras in CONTRIBUTING.md.

* :white_check_mark: Unit test to create datashader.Canvas from xr.Dataset

Ensure that xarray.Dataset objects can be turned into a datashader.Canvas too! Clarified that the `.rio.x_dim` and `.rio.y_dim` attributes are required, and made some minor stylistic changes to the docstring and code in datashader.py.

* :rotating_light: Remove type hint for __iter__ in XarrayCanvas class

Fixes `AttributeError: 'NoneType' object has no attribute 'Canvas'` when the optional datashader dependency is not installed.