:art: Ensure that unsupported GeometryCollection types error gracefully
Improved traceback error and added a unit test for when GeometryCollection vector types (i.e. those with an assortment of point, line or polygon types) are passed in to DatashaderRasterizer. Limitation is on spatialpandas really, and hence datashader. Also fallback having datashader.Canvas's CRS to None to prevent AttributeError, though that might cause some issues when the vector has a CRS but the canvas doesn't.