:memo: Document default aggregation function and error when missing crs
Improving the DatashaderRasterizer docstring so that people know what is happening. Mention that the default aggregation is 'count' for points, and 'any' for lines and polygons. Document AttributeError that is raised when either the canvas or vector input is missing a `.crs` attribute, and ValueError raised when vector geometry type is not supported. Also added an intersphinx link for shapely.