:mag: Show more levels for the in-page table of contents (#36)
Making it easier to navigate to the desired subsections, especially on the API reference page.

* :mag: Add table of contents to top of API reference page

Make it easier to jump to the relevant sections, since the list of DataPipes is increasing!

* :memo: Order DataPipe subsections on API page alphabetically

Swapping positions of Pyogrio and Rioxarray datapipes.

* :zap: Cache execution outputs when building jupyter book docs

No need to build pages again if things haven't changed, see https://jupyterbook.org/en/latest/content/execute.html.

* :alien: Increase depth of secondary sidebar on the right from 2 to 3

Found the right sphinx-book-theme config setting to increase the depth of the secondary sidebar (on the right) from 2 to 3. See https://sphinx-book-theme.readthedocs.io/en/stable/customize/sidebar-secondary.html. Decided too that the in-page table of contents can go.