:fire: Focus on just one vector shapefile instead of two
Marie Kondo-ing the section on pre-processing the vector data, because working with two shapefiles was a pain. The merging of two GeoDataFrame objects using BatchMapper was a bad idea as it returned a DataPipe without a length, so DatashaderRasterizer complained. Decided too that focusing on just one vector was enough to teach the concept, and the masters can read the footnotes on more advanced combinatorial pipelines.