:memo: Writeup section on combining Sentinel-1 image and water mask
Combining two xarray.DataArray objects into a single xarray.Dataset before slicing. Using Zipper to pair the two DataPipes, and then Collator to do the actual combining. Showing the DataPipe graph after the zip stage to give people an overview of the (many) steps done so far. Double checked too that the image and mask looks ok, which led to the discovery of an issue with rounding numbers or something. Had to change the resolution from 80m to 100m because of this, otherwise the rasterized water mask is cut off at the Southern part.