:beers: Finalize tutorial by loading tensors into dataloader
Just complete the whole vector segmentation tutorial in one go! Really simplified the ending with a minimal xbatcher slicing and conversion to torch.Tensor step only. Gave a shoutout to Edoardo's work at UNOSAT in the end. Decided to rename the file to vector-segmentation-masks. Added more emojis throughout the page, and should be good to go!