:dizzy: Reproject vector to match raster's coordinate reference system
To handle raster images with different coordinate reference systems, do on-the-fly reprojection of the vector geopandas.GeoDataFrame to match the raster xarray.DataArray's coordinate reference system. Have added a unit test with EPSG:3857 and EPSG:32613 chips, and noted this automatic reprojection feature in the docstring. Need to refactor this basic implementation to be more efficient as the whole geodataframe is being reprojected rather than a subset only.