:sparkles: GeoPandasRectangleClipper for spatially subsetting vectors (#52)
An iterable-style DataPipe for clipping vector geometries with the rectangular bounding box extent of raster grids!

* :sparkles: GeoPandasRectangleClipper for spatially subsetting vectors

An iterable-style DataPipe for clipping vector geometries with a raster grid! Uses the rectangular clip algorithm in geopandas.clip for the spatial subsetting. Included a doctest that ensures a geopandas.GeoDataFrame can be clipped by two xarray.DataArray grids. Added a new section in the API docs too.

* :white_check_mark: Unit tests to clip geoseries and check incorrect len

Ensure that a geopandas.Geoseries vector can be clipped using xarray.Dataset rasters, and that vector datapipes which are too long (len > 1) raise a NotImplementedError properly.

* :dizzy: Reproject vector to match raster's coordinate reference system

To handle raster images with different coordinate reference systems, do on-the-fly reprojection of the vector geopandas.GeoDataFrame to match the raster xarray.DataArray's coordinate reference system. Have added a unit test with EPSG:3857 and EPSG:32613 chips, and noted this automatic reprojection feature in the docstring. Need to refactor this basic implementation to be more efficient as the whole geodataframe is being reprojected rather than a subset only.