:speech_balloon: Edit type hints, isort imports and quick docstring fixes (#54)
Just a random collection of mostly documentation-related patches. Patches type-hints in #52, isort imports in #35, mention functional name of IterDataPipe in walkthroughs #8 and #20, and remove mention of returned tuple to patch #33.

* :label: Add specific type hints for mask_datapipe in geopandas.py

Should be either an xarray.DataArray or xarray.Dataset.

* :rotating_light: Sort spatialpandas imports in datashader.py

Ran isort to sort spatialpandas.geometry imports alphabetically. Also intersphinx linked the `.crs` attribute to geopandas.GeoDataFrame.crs.

* :speech_balloon: Mention functional name of IterDataPipe in walkthroughs

So people don't get confused on why the class-form like `Collator` is mentioned but `.collate` was used instead.

* :memo: Remove mention of tuple being returned in test_pyogrio_reader

Forgot to edit the unit test's docstring. Patches #33.

* :beers: It's GeoPackage and GeoDataFrame, not GeoTIFF and DataArray

Need to be more careful when copying and pasting stuff.