:sparkles: PySTACItemReaderIterDataPipe for reading STAC Items (#46)
An iterable-style DataPipe for STAC items! Uses pystac for reading the files or URLs into pystac.Item objects.

* :sparkles: PySTACItemReaderIterDataPipe for reading STAC Items

An iterable-style DataPipe for STAC items! Uses pystac for reading the files or URLs. Included a doctest that checks the metadata within the pystac.item.Item object. Added a new section in the API docs and an intersphinx mapping.

* :green_heart: Make pystac an extras dependency in the spatial category

Ensure that zen3geo works even when `pystac` is not installed and add `pystac` to the spatial section of the extras dependencies in pyproject.toml.

* :truck: Rename functional form as read_to_pystac_item

Decided that since the returned object is a `pystac.Item`, it should probably be `read_to_pystac_item`.

* :white_check_mark: Unit test to read JSON and return a pystac.Item

Ensure that a JSON STAC item can be read into a pystac.Item object that contains various spatiotemporal metadata.

* :green_heart: Skip PySTACItemReader doctests if pystac not installed

Use pytest.importorskip to skip running the doctest when pystac cannot be imported.