:sparkles: PySTACAPISearchIterDataPipe to query dynamic STAC Catalogs (#59)
An iterable-style DataPipe to make STAC API queries to dynamic STAC Catalogs! Uses pystac-client to perform the STAC API Item Search on the `/search` endpoint.

* :heavy_plus_sign: Add pystac-client

Python client for STAC Catalogs and APIs!

* :sparkles: PySTACAPISearchIterDataPipe for searching STAC Catalogs

An iterable-style DataPipe to make STAC API queries! Uses pystac-client to perform the STAC API Item Search. Included a doctest that does a STAC API search which returns an iterable list of STAC item objects. Added a new section in the API docs and an intersphinx mapping.

* :white_check_mark: Unit test to do STAC query and return ItemSearch

Ensure that a spatiotemporal query can be made to a STAC API /search/ endpoint to produce an instance of a pystac_client.ItemSearch object containing pointers to multiple pystac.Item objects. Also removed a for-loop that was accidentally added after a yield statement.

* :recycle: Reserve kwargs for Client.open instead of Client.search

Decided that it is cleaner to pass in all the STAC API query parameters via the source_datapipe, rather than through both the source_datapipe and kwargs. The kwargs can then be kept for keyword arguments into pystac.Client.open. Documented some example parameters for pystac_client.Client.search and pystac_client.Client.open. Updated the unit test and doctest, adding a teaser way to sign planetary_computer assets inplace too!

* :white_check_mark: Unit test to set pystac_client.Client.open parameters

Ensure that parameters such as authentication keys can be passed to pystac_client.Client.open. Using a fake API key for Radiant MLHub, which can only take us so far as to getting a title and description.

* :green_heart: Remove client validate because jsonschema not installed

Fix Continuous Integration failure because jsonschema had to be installed but it is an optional dependency.

* :package: Create new extras dependency section for the STAC stack

Move pystac and pystac-client to a dedicated `stac` extras section for spatiotemporal dependencies! Also updated index page with instructions on installing these STAC dependencies.