:sparkles: StackSTACStackerIterDataPipe for stacking STAC items
An iterable-style DataPipe to stack STAC items! Uses stackstac to do the reprojection and stacking along time. Included a doctest that stacks the VH and VV channels of a Sentinel-1 image into an xarray.DataArray. Added a new section in the API docs and intersphinx mappings for dask and stackstac.