:memo: Writeup section on loading polygons and stacking steps and params
Show how to get the vector layer loaded as a geopandas.GeoDataFrame and added an overview of the layer stacking steps. Reduced the size of the Sentinel-1 xarray.DataArray by clipping to just Gunung Talamau, reducing the spatial resolution from 10m to 30m, and optimizing dtype from float64 to float16. Also added link to Humanitarian Data Exchange where the nice link to the shapefile (on a server that supports HTTP range requests) can be found.