:sparkles: StackSTACMosaicIterDataPipe to mosaic tiles into one piece
An iterable-style DataPipe to mosaic tiles in a DataArray! Uses stackstac to do the mosaic operation. Included a doctest that mosaicks two ALOS DEM tiles (stored at different positions in the time index of an xarray.DataArray) into an xarray.DataArray with dimensions (band, y, x). Added a new section in the API docs too!