:sparkles: Implement len function for XbatcherSlicerIterDataPipe (#75)
* :sparkles: Implement len function for XbatcherSlicerIterDataPipe

Allow the number of chips from XbatcherSlicer datapipes to be calculated directly using `len()`! Not previously implemented because it was inefficient in xbatcher, but xbatcher v0.2.0 has implemented lazy batch generation so this should be performant enough now. Also fixed an incorrect intersphinx link to xbatcher.BatchGenerator in the docstring.

* :memo: Get len of datapipes directly instead of using list comprehension

Previously turned the datapipes into a list to get the length, now it's possible to avoid the intermediate list representation!