:bug: Fix DatashaderRasterizer to allow N:1 instead of just 1:1 (#98)
* :bug: Fix DatashaderRasterizer to allow N:1 instead of just 1:1

Used to wrong boolean operator (should be and, not or) which meant DatashaderRasterizer didn't work when the vector_datapipe was length 1 and canvas_datapipe was length 2 or more. This prevents false errors like `ValueError: Unmatched lengths for the canvas datapipe (XarrayCanvasIterDataPipe) and vector datapipe (PyogrioReaderIterDataPipe). The vector datapipe's length (1) should either be (1) to allow for broadcasting, or match the canvas datapipe's length of (2)`.

Patches the bugfix at #39 from zen3geo v0.3.0. Also updated a unit test to use a 2:1 canvas:vector ratio to prevent regression.

* :pushpin: Temporarily pin numpy<1.24 to prevent AttributeError

The latest numpy=1.24 version installed via readthedocs is incompatible with datashader=0.14.4, resulting in an error like `AttributeError: module 'numpy' has no attribute 'warnings'`. This was reported in https://github.com/holoviz/datashader/issues/1209 and fixed in https://github.com/holoviz/datashader/pull/1176, but will need to wait for datashader>=0.14.5 to be released.